spec tests overhaul: - refactor User model tests - create Channel model tests - move db test into initializers - run db tests before all others (so models have access to mongo)
